Move binary conversion to separate class
